Camp fixes

+++ src/main/java/krispasi/omGames/hallsofcarnage/HallsBreakableType.java
@@ -17,9 +17,9 @@ public record HallsBreakableType(String id,
         loot = List.copyOf(loot);
     }
 
-    public record Part(int offsetX,
-                       int offsetY,
-                       int offsetZ,
+    public record Part(double offsetX,
+                       double offsetY,
+                       double offsetZ,
                        Material material,
                        String blockData,
                        double rotationX,
